Allow executeToResultPane to take "finished" callback
No, I couldn't manage to make this type with just `typedef`; I had to use functional. And no, I can't manage to change the action function type to use functional either for consistency's sake and I don't know why. So I'm using two different ways to define a function pointer type and we all have to live with it.

AND I have to bump the C++ standard from 11 to 17 to declare a static const lambda in the header file. I guess you just couldn't do this in C++11???