Switch User: Save last username
Will now automatically prompt you to log in as the last domain\username you entered to Switch User on launch. If you cancel the prompt or have never switched user before, it will just start as yourself as normal.

Still looking for a way to have Windows save the password for me too, because that would be nice.